Move QAttention/QEmbedLayerNormalization op defs to quantization_defs.cc (#10507)
