+++ onnxruntime/core/graph/contrib_ops/shape_inference_functions.h
@@ -1,0 +1,17 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+// This file contains some helper functions that are used for implementing ONNX type/shape inference.
+
+namespace ONNX_NAMESPACE {
+struct InferenceContext;
+}
+
+namespace onnxruntime {
+namespace contrib {
+void AttentionTypeAndShapeInference(ONNX_NAMESPACE::InferenceContext& ctx, int past_input_index);
+void EmbedLayerNormalizationShapeInference(::ONNX_NAMESPACE::InferenceContext& ctx);
+}
+}  // namespace onnxruntime
